@@ -2,10 +2,10 @@
   "name": "tinper-sparrow",
   "version": "3.1.27",
   "description": "sparrow.js",
-  "main": "js/index.js",
+  "main": "src/index.js",
   "scripts": {
-    "build": "webpack --colors --progress --mode=build",
-    "dev": "webpack --progress --colors  --mode=dev",
+    "build": "webpack --colors --progress --env.mode=build",
+    "dev": "webpack --progress --colors --env.mode=dev",
     "test": "mocha --compilers js:babel-core/register --colors -w ./test/*.spec.js",
     "es": "babel src -d lib",
     "prod": "npm run es && npm run build",
@@ -33,12 +33,13 @@
     "babel-cli": "^6.11.4",
     "babel-core": "^6.11.4",
     "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
+    "babel-polyfill": "^6.3.14",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.3.13",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.11.5",
     "babel-plugin-transform-es3-member-expression-literals": "^6.8.0",
     "babel-plugin-transform-es3-property-literals": "^6.8.0",
-    "babel-preset-es2015": "^6.9.0",
     "babel-preset-es2015-loose": "^8.0.0",
     "babel-preset-stage-0": "^6.5.0",
     "chai": "^3.5.0",
@@ -47,7 +48,7 @@
     "file": "^0.2.2",
     "fs-extra": "^2.0.0",
     "mocha": "^2.5.3",
-    "webpack": "^1.13.1",
+    "webpack": "^2.2.1",
     "yargs": "^4.8.1"
   }
 }
